complete exclusion criterion; assorted bug fixes
The exclusion criterion which attempts to determine when calculator
annotations are are complete and reliable has been completed.
This criterion largely boiled down to the size of two sets:
`unspecified_or_underived` or `unused`.
`unspecified_or_underived` is the set of quantities which appear in
text steps but were not extracted from the answer text, and which do not appear in the right
hand side of any annotated calculation (in which case we could assume
that the quantity is derived and *should* not be expected in the question
text).  Thus, quantities in `unspecified_or_underived` either were not detected in the question
when they should have been, or are on the rhs of some missing
calculator annotation.  `unused` is the set of quantities which are
present in both the question and the text steps, but which do not appear
on the lhs of any calculator annotation.  This suggests that a quantity
provided by the question and referenced in the text steps was not used
to derive a quantity via a calculator annotation.  Intuitively, in most
cases this is because the calculator annotation utilizing those inputs
is missing.  Thus, if `unused` is nonempty, it is likely that there is
simply a missing calculator annotation.
If `unspecified_or_underived` is nonempty, perhaps this signals an
issue with question quantity detection that
can be rectified, or perhaps calculator annotations are incomplete.
It's difficult to say.

In addition to this, several bugs were corrected including:
 - `INT_OR_FLT` was modified to allow hanging decimals
 - `CALCULATION` was broken up into a regex for extracting whole
   annotations, and a regex for extracting the equation inside of an
   annotation.
 - A new regex, `HYPHENATION`, was added for use in dehyphenation;
   calls to `*_raw.replace("-", "")` were also removing minus signs
 - `decompose_question` was refactored to correct the issue that numeral
   parsing was not being triggered in the case where a numeral is
   directly followed by a digitized quantity.
 - `decompose_answer` was modified to provide quantities in the lhs and
   rhs of calculator annotations separately.